pyproject.toml
[project]
name = "bytewatt-export-limiter"
version = "1.0.0"
description = "Home Assistant integration for Bytewatt battery export limiting with price-based automation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Carlo"}
]
keywords = ["home-assistant", "bytewatt", "modbus", "solar", "export-limiting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pymodbus>=3.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/carlo/bytewatt-export-limiter"
Documentation = "https://github.com/carlo/bytewatt-export-limiter#readme"
Repository = "https://github.com/carlo/bytewatt-export-limiter"
Issues = "https://github.com/carlo/bytewatt-export-limiter/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["custom_components*"]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
]
ignore = [
    "PLR0913",  # Too many arguments to function call
    "PLR2004",  # Magic value used in comparison
    "PLR1714",  # Multiple comparisons can use set - less readable for our case
    "E501",     # Line too long (handled by formatter)
    "TC001",    # Type checking blocks - overkill for HA integrations
    "TC002",    # Type checking blocks - overkill for HA integrations
    "TC003",    # Type checking blocks - overkill for HA integrations
    "TCH001",   # Type checking blocks - overkill for HA integrations
    "TCH002",   # Type checking blocks - overkill for HA integrations
    "TCH003",   # Type checking blocks - overkill for HA integrations
    "ARG001",   # Unused function argument - common in HA interface implementations
    "ARG002",   # Unused method argument - common in HA interface implementations
    "ERA001",   # Commented-out code - we have intentional TODO comments
    "PLC0415",  # Import not at top of file - needed for lazy imports in HA
    "SIM102",   # Nested if - sometimes clearer to keep separate
    "SIM105",   # contextlib.suppress - try/except is clearer for async cancellation
    "ASYNC109", # Timeout parameter naming - false positive for our use case
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "ARG",      # Unused arguments in tests (fixtures)
    "PLR2004",  # Magic values in tests are fine
    "PLC0415",  # Inline imports in tests are common pattern
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "-v",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

# Coverage configuration
[tool.coverage.run]
source = ["custom_components/bytewatt_export_limiter"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 70
show_missing = true

# Mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
